menyelesaikan fitur kelas

+++ backend/auth_service/models/user.go
@@ -2,11 +2,11 @@ package models
 
 type User struct {
 	// gorm.Model
-	ID        int32  `gorm:"primaryKey"`
+	ID        uint64 `gorm:"primaryKey"`
 	Username  string `gorm:"unique;not null"`
 	Email     string `gorm:"unique;not null" json:"email" binding:"required,email"`
 	Password  string `gorm:"not null"`
 	Role      string
-	SekolahID int32
+	SekolahID uint32
 	Sekolah   Sekolah `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 }
